Change command names, remove dev_command file.
